refactor(tests): consolidate and reorganize test files (#3616)
